requirements.txt
fastapi
uvicorn
openai
pydantic
python-multipart
streamlit
requests
python-dotenv
pyttsx3
reportlab
transformers
torch
gTTS==2.5.1
deep-translator
python-jose[cryptography]
requests
fastapi[all] 
httpx
sqlalchemy 
psycopg2-binary
